core: Normalize the use of the e-mail closing across all e-mail templates
There is now a single {% block closing %} which contains the text
"Kind regards, <sitename>".

This is done to make it easier to override this closing with a
different one across all e-mails for a site.

Note that the password reset and email verification mails used a
different closing, ("Thank you, the team at ..."); that closing has
been replaced with the "kind regards" one because of this
normalisation.