mod_base: Fix remember password for Chrome
Added class "z_logon_form" to the logon form, and let its target be a
hidden iframe. When a z_logon_form is posted, we post to this hidden
iframe, to trick chrome into showing the "save password" option. The
logon form submit event cannot be cancelled because that blocks the
form-remembering.

Fixes #491