core: catinclude also accepts a list of atoms instead of an id
This makes it in the admin again possible to override the
`_admin_overview_list.tpl` file based on the category that is
selected. This restores the functionality that was removed in commit
1d715afeef5a6823dd0bb38cbaee8f41497a9345, but now implemented in a
different, more explicit way.

Documentation for the new `m.category[222].is_a` option and the extra
feature of `catinclude` is provided.